refactor: merge schemas CharacterIngredients and LightConeIngredients into Ingredients
