+++ packages/nuxt/types/data/ingredient.ts
