--- packages/nuxt/types/data/src/equipment-ingredients.ts
@@ -1,11 +1,0 @@
-import type { EachLevels, Ingredient } from "../ingredients"
-
-export interface EquipmentIngredients {
-  expItems: {
-    itemId: string
-    expPerItem: number
-  }[]
-  levelingItemTables: Record<string, IngredientTable>
-}
-
-type IngredientTable = EachLevels<Ingredient[]>
