+++ packages/nuxt/types/data/src/equipments.ts
@@ -1,12 +1,12 @@
 import type { HsrPath } from "../enums"
-import type { MaterialExpr } from "../ingredients"
+import type { MaterialExpr } from "../ingredient"
 
 interface EquipmentBase {
   id: string
   $nameJA: string
   yomi: string
   rarity: number
-  materials?: Record<string, MaterialExpr>
+  materials: Record<string, MaterialExpr>
 }
 
 export interface LightCone extends EquipmentBase {
